podidentity/requirements.txt
azure-servicebus
azure-identity
requests